Add files via upload
This dataset presents the number of electric vehicles registered in the province of Nova Scotia, Canada, with an active licence plate attached. 

This data breaks down the number of electric vehicles by vehicle model year, vehicle make and County of Nova Scotia in which the licence plate owner resides.

Try visualizing this dataset using any tool you like. The main goal is to create a visualization that highlights useful insights from the data.